backend/requirements-prod.txt
fastapi
# supabase
openai>=1.0.0
uvicorn[standard]
termcolor
tiktoken
numpy
python-dotenv
Requests
markdown
APScheduler
pytz
pydantic[email]>=2.4.2
python-multipart
# pytest
# pytest-asyncio
# pytest-mock
# pytest-cov
promptmodel>=0.1.0
sqlalchemy>=2.0.0
python-dotenv
alembic
asyncpg
fastapi_nextauth_jwt
passlib
bcrypt==4.0.1
psycopg2-binary
greenlet
redis
PyJWT